# openapi.yaml
openapi: 3.1.0
info:
  title: BehaviorIQ Assessment API
  description: |
    API for behavioral assessment trials and full assessments with Stripe payment integration.
    This API powers the BehaviorIQ ChatGPT App for taking behavioral assessments directly in ChatGPT.
  version: 1.0.0
  contact:
    name: BehaviorIQ Support
    url: https://behavioriq.com/support
    email: support@behavioriq.com
  x-logo:
    url: https://behavioriq.com/logo.png
servers:
  - url: https://app.behavioriq.com
    description: Production API
  - url: http://localhost:3000
    description: Development API
security:
  - ApiKeyAuth: []
paths:
...
  /api/user/credits:
    get:
      operationId: getUserCredits
      summary: Get user's remaining credits
      description: |
        Returns the number of remaining assessment credits for the authenticated user.
        Requires API key authentication.
      parameters:
        - name: user_id
          in: query
          description: User ID to check credits for
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User credits retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditsResponse"
        "401":
          description: Unauthorized - invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - User
...
  /api/assessment/start:
    post:
      operationId: startFullAssessment
      summary: Start a full assessment
      description: |
        Initiates a full 75-question assessment. Requires at least 1 credit.
        If user has no credits, returns checkout URL for purchasing credits.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: Authenticated user ID
                childName:
                  type: string
                  description: Name of the child being assessed
                childAge:
                  type: integer
                  minimum: 3
                  maximum: 18
                  description: Age of the child
                relationshipType:
                  type: string
                  enum:
                    - parent
                    - educator
                    - other
                  description: Relationship of respondent to child
              required:
                - userId
                - childName
                - childAge
                - relationshipType
      responses:
        "200":
          description: Full assessment started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentStartResponse"
        "402":
          description: Insufficient credits - checkout required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsufficientCreditsResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - Full Assessment
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Get yours from the dashboard.
  schemas:
...
    CreditsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        userId:
          type: string
        credits:
          type: integer
          minimum: 0
          example: 3
        plan:
          type: string
          enum:
            - free
            - core_monthly
            - family_monthly
            - core_annual
            - family_annual
          example: "core_monthly"
        nextCreditDate:
          type: string
          format: date-time
          description: Date when next credit batch will be granted
          example: "2025-12-03T00:00:00Z"
        creditsCap:
          type: integer
          description: Maximum credits that can roll over
          example: 6
      required:
        - success
        - userId
        - credits
        - plan
...
    AssessmentStartResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        assessmentId:
          type: string
          description: ID for this assessment session
          example: "assess_123abc"
        childName:
          type: string
          example: "Emma"
        childAge:
          type: integer
          example: 8
        questions:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentQuestion"
          description: Array of 75 assessment questions
        totalQuestions:
          type: integer
          example: 75
        estimatedTime:
          type: string
          example: "20-30 minutes"
        creditDeducted:
          type: integer
          example: 1
        creditsRemaining:
          type: integer
          example: 2
      required:
        - success
        - assessmentId
        - questions
        - totalQuestions
...
    InsufficientCreditsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "insufficient_credits"
        message:
          type: string
          example: "You need 1 credit to start an assessment. Purchase credits to continue."
        creditsNeeded:
          type: integer
          example: 1
        creditsAvailable:
          type: integer
          example: 0
        checkoutUrl:
          type: string
          format: uri
          description: Checkout URL to purchase credits
          example: "https://checkout.stripe.com/pay/cs_live_abc123xyz"
      required:
        - success
        - error
        - message
        - checkoutUrl
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error code/type
          example: "invalid_parameters"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid child age. Must be between 3 and 18."
      required:
        - success
        - error
        - message

    # Supporting Schemas
...
    AssessmentQuestion:
      type: object
      properties:
        id:
          type: string
          example: "q_1"
        text:
          type: string
          example: "Difficulty sustaining attention to tasks or play"
        domain:
          type: string
          example: "attention"
        category:
          type: string
          example: "Attention & Focus"
        options:
          type: array
          items:
            type: string
          example:
            - never
            - rarely
            - sometimes
            - often
            - very_often
      required:
        - id
        - text
        - domain
        - category
        - options
...